+++ tests/components/tts/test_init.py
@@ -1,6 +1,5 @@
 """The tests for the TTS component."""
 import asyncio
-from collections.abc import Generator
 from http import HTTPStatus
 from typing import Any
 from unittest.mock import MagicMock, patch
@@ -19,11 +18,10 @@ from homeassistant.components.media_player import (
 )
 from homeassistant.components.media_source import Unresolvable
 from homeassistant.components.tts.legacy import _valid_base_url
-from homeassistant.config_entries import ConfigEntry, ConfigEntryState, ConfigFlow
+from homeassistant.config_entries import ConfigEntryState
 from homeassistant.const import ATTR_ENTITY_ID, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant, State
 from homeassistant.exceptions import HomeAssistantError
-from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.setup import async_setup_component
 import homeassistant.util.dt as dt_util
 from homeassistant.util.network import normalize_url
@@ -31,122 +29,24 @@ from homeassistant.util.network import normalize_url
 from .common import (
     DEFAULT_LANG,
     SUPPORT_LANGUAGES,
+    TEST_DOMAIN,
     MockProvider,
-    MockTTS,
     MockTTSEntity,
     get_media_source_url,
+    mock_config_entry_setup,
+    mock_setup,
 )
 
-from tests.common import (
-    MockConfigEntry,
-    MockModule,
-    MockPlatform,
-    async_mock_service,
-    mock_config_flow,
-    mock_integration,
-    mock_platform,
-    mock_restore_cache,
-)
+from tests.common import async_mock_service, mock_restore_cache
 from tests.typing import ClientSessionGenerator, WebSocketGenerator
 
 ORIG_WRITE_TAGS = tts.SpeechManager.write_tags
-TEST_DOMAIN = "test"
 
 
 @pytest.fixture
 async def setup_tts(hass: HomeAssistant, mock_tts: None) -> None:
     """Mock TTS."""
     assert await async_setup_component(hass, tts.DOMAIN, {"tts": {"platform": "test"}})
-
-
-class TTSFlow(ConfigFlow):
-    """Test flow."""
-
-
-@pytest.fixture(autouse=True)
-def config_flow_fixture(hass: HomeAssistant) -> Generator[None, None, None]:
-    """Mock config flow."""
-    mock_platform(hass, f"{TEST_DOMAIN}.config_flow")
-
-    with mock_config_flow(TEST_DOMAIN, TTSFlow):
-        yield
-
-
-@pytest.fixture(name="setup")
-async def setup_fixture(
-    hass: HomeAssistant,
-    request: pytest.FixtureRequest,
-    mock_provider: MockProvider,
-    mock_tts_entity: MockTTSEntity,
-) -> None:
-    """Set up the test environment."""
-    if request.param == "mock_setup":
-        await mock_setup(hass, mock_provider)
-    elif request.param == "mock_config_entry_setup":
-        await mock_config_entry_setup(hass, mock_tts_entity)
-    else:
-        raise RuntimeError("Invalid setup fixture")
-
-
-async def mock_setup(
-    hass: HomeAssistant,
-    mock_provider: MockProvider,
-) -> None:
-    """Set up a test provider."""
-    mock_integration(hass, MockModule(domain=TEST_DOMAIN))
-    mock_platform(hass, f"{TEST_DOMAIN}.{tts.DOMAIN}", MockTTS(mock_provider))
-
-    await async_setup_component(
-        hass, tts.DOMAIN, {tts.DOMAIN: {"platform": TEST_DOMAIN}}
-    )
-    await hass.async_block_till_done()
-
-
-async def mock_config_entry_setup(
-    hass: HomeAssistant, tts_entity: MockTTSEntity
-) -> MockConfigEntry:
-    """Set up a test tts platform via config entry."""
-
-    async def async_setup_entry_init(
-        hass: HomeAssistant, config_entry: ConfigEntry
-    ) -> bool:
-        """Set up test config entry."""
-        await hass.config_entries.async_forward_entry_setup(config_entry, tts.DOMAIN)
-        return True
-
-    async def async_unload_entry_init(
-        hass: HomeAssistant, config_entry: ConfigEntry
-    ) -> bool:
-        """Unload up test config entry."""
-        await hass.config_entries.async_forward_entry_unload(config_entry, tts.DOMAIN)
-        return True
-
-    mock_integration(
-        hass,
-        MockModule(
-            TEST_DOMAIN,
-            async_setup_entry=async_setup_entry_init,
-            async_unload_entry=async_unload_entry_init,
-        ),
-    )
-
-    async def async_setup_entry_platform(
-        hass: HomeAssistant,
-        config_entry: ConfigEntry,
-        async_add_entities: AddEntitiesCallback,
-    ) -> None:
-        """Set up test tts platform via config entry."""
-        async_add_entities([tts_entity])
-
-    loaded_platform = MockPlatform(async_setup_entry=async_setup_entry_platform)
-    mock_platform(hass, f"{TEST_DOMAIN}.{tts.DOMAIN}", loaded_platform)
-
-    config_entry = MockConfigEntry(domain=TEST_DOMAIN)
-    config_entry.add_to_hass(hass)
-    assert await hass.config_entries.async_setup(config_entry.entry_id)
-    await hass.async_block_till_done()
-
-    return config_entry
 
 
 class DefaultEntity(tts.TextToSpeechEntity):
